backend/requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
gunicorn==21.2.0
psycopg[binary]==3.2.9
bcrypt==4.0.1
flask-jwt-extended==4.6.0
resend==2.11.0